Add implementation for BankTransaction class with immutable properties.
The properties are immutable of this class, as they are designed to not be changed after the object is created. This is
done by using automatic properties without a set accessor, making them only modifiable in the constructor. This could
also be done with readonly fields, but automatic properties are more concise and easier to read.