Add implementation of the BankingSystem class.
The BankingSystem class is used to manage the bank accounts and bank transactions. It keeps a record of all
account objects, along with their associated transactions. It has functions for printing lists of accounts and
transactions, as well as creating new transactions on accounts. The other component classes were modified, adding
overrides to the ToString() method to allow for better printing of the objects.

Other small changes include the BankAccount withdraw method returning a bool to indicate success. Name was also
renamed in this class to OwnerName to better reflect the data it holds. Finally, some test code was added to the main
Program class to test the new functionality of the BankingSystem class.